Rename helper and add logic to detect janus path
